% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmur.R
\name{dmur}
\alias{dmur}
\title{Statistical Decision Analysis: with the help of this package we construct the payoff table and the opportunity loss table and calculate various statistical parameters like EMVs, MEMV, EPPI, EVPI, EOL which help in taking effective decisions.}
\usage{
dmur(dAmount, prob, sPrice, pPrice)
}
\arguments{
\item{dAmount}{A data vector to store values}

\item{prob}{A data vector to store the probabilities}

\item{sPrice}{A numeric value to store selling price}

\item{pPrice}{A numeric value to store preparation/ purchase price}
}
\value{
MEMV, bestValue, EPPI, EVPI, EOL
}
\description{
Statistical Decision Analysis: with the help of this package we construct the payoff table and the opportunity loss table and calculate various statistical parameters like EMVs, MEMV, EPPI, EVPI, EOL which help in taking effective decisions.
}
\examples{
dmur(c(50, 100, 150, 200), c(0.2,0.4,0.3,0.1), 20, 10)

}
\references{
1. Khalili, k., Damghani, M. T., Taghavifard, R., Tavakkoli M., (2009) Decision making under uncertain and risky situations, Enterprise risk management symposium monograph society of Actuaries- Schaumburg, Illinois, vol. 15.

2. Marakas, G. M. (2006) Decision support system- In the 21st century, 2nd edition, Pearson education, New Delhi.

3. Turban, E., Aronson, J. E. and Liang, T.P. (2006) Decision support system and intelligent systems, Pearson education, New Delhi.
}

